backend/pyproject.toml
[project]
name = "flask-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "blinker>=1.6.2",
    "click>=8.1.3",
    "colorama>=0.4.6",
    "fastapi>=0.115.12",
    "firebase-admin>=6.6.0",
    "flask-cors>=5.0.0",
    "flask-migrate>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gspread>=6.1.4",
    "httpx>=0.28.1",
    "itsdangerous>=2.1.2",
    "jinja2>=3.1.2",
    "markupsafe>=2.1.2",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "oauth2client>=4.1.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pyftdi>=0.54.0",
    "pyserial>=3.5",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pyusb>=1.2.1",
    "pyyaml>=6.0",
    "rich>=14.0.0",
    "schedule>=1.2.2",
    "scipy>=1.15.1",
    "setuptools>=58.0.4",
    "six>=1.16.0",
    "uvicorn>=0.34.2",
    "waitress>=3.0.2",
    "werkzeug>=2.3.4",
    "zipp>=3.15.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "rich>=14.0.0",
    "ruff>=0.11.12",
]
# prod = [
#     "picamera2>=0.3.25; sys_platform == 'linux'"
# ]
